Move ws2812_common.h to the drivers directory
