+++ drivers/avr/ws2812.h
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "quantum/color.h"
-#include "quantum/ws2812_common.h"
+#include "drivers/ws2812_common.h"
 
 /* User Interface
  *
