+++ drivers/ws2812_common.h
